Backend/requirements.txt
flask
flask-cors
python-docx
PyPDF2
groq
mammoth
tensorflow
pandas
scikit-learn
matplotlib